requirements.txt
pandas>=1.3.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.0.0
python-dotenv>=0.19.0
matplotlib>=3.4.0
seaborn>=0.11.0
